+++ src/main/java/toolc/yourlist/auth/domain/AllPasswordPolicy.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.auth.domain.request;
+package toolc.yourlist.auth.domain;
 
 class AllPasswordPolicy implements PasswordPolicy {
