+++ src/main/java/toolc/yourlist/auth/domain/AllowOnlyAlphabetOrNumberOrSpecialCharacter.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.auth.domain.request;
+package toolc.yourlist.auth.domain;
 
 import java.util.Arrays;
