+++ src/main/java/toolc/yourlist/auth/domain/MemberLogin.java
@@ -2,7 +2,6 @@ package toolc.yourlist.auth.domain;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import toolc.yourlist.auth.domain.request.LoginRequest;
 import toolc.yourlist.member.domain.AllMember;
 
 @RequiredArgsConstructor
