+++ src/main/java/toolc/yourlist/auth/domain/NonNull.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.auth.domain.request;
+package toolc.yourlist.auth.domain;
 
 class NonNull implements LoginIdPolicy {
