+++ src/main/java/toolc/yourlist/auth/domain/Password.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.auth.domain.request;
+package toolc.yourlist.auth.domain;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
