+++ src/main/java/toolc/yourlist/auth/domain/PasswordNonNull.java
@@ -1,4 +1,4 @@
-package toolc.yourlist.auth.domain.request;
+package toolc.yourlist.auth.domain;
 
 class PasswordNonNull implements PasswordPolicy {
