+++ src/test/java/toolc/yourlist/auth/domain/LoginIdTest.java
@@ -1,6 +1,7 @@
-package toolc.yourlist.auth.domain.request;
+package toolc.yourlist.auth.domain;
 
 import org.junit.jupiter.api.Test;
+import toolc.yourlist.auth.domain.LoginId;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
