+++ src/test/java/toolc/yourlist/auth/domain/PasswordPolicyTest.java
@@ -1,10 +1,10 @@
-package toolc.yourlist.auth.domain.request;
+package toolc.yourlist.auth.domain;
 
 import org.junit.jupiter.api.Test;
+import toolc.yourlist.auth.domain.*;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static org.junit.jupiter.api.Assertions.*;
 
 class PasswordPolicyTest {
